Update user data when user changes names
When a user changes his/her Flowdock account Nick, the only way to update the user data is to manually modify the users object. This affects scripts such as roles, karma, etc. that attach data to user objects based on the associated names. These changes add a check for the 'user-edit' stream event and remaps the user object in the brain.