Skipping the users who left a flow
This change adds a logic to filter out the users who left a flow. Such users can be determined by the "in_flow" flag set to "false". The thing is that flowdock allows several users in a flow to have the same nick, whereas hubot treats a nickname as a unique entity. The change doesn't fix completely the problem, but at least reduce it.